remove default font size, make font sizes floats, minor style improvements

+++ core/src/commonMain/kotlin/warlockfe/warlock3/core/prefs/HighlightRepository.kt
@@ -132,7 +132,7 @@ class HighlightRepository(
             italic: Boolean,
             underline: Boolean,
             fontFamily: String?,
-            fontSize: Double? ->
+            fontSize: Float? ->
             Pair(
                 groupNumber,
                 StyleDefinition(
